Renaming: SessionReceiver -> MessageSession,  SessionReceiverOuter -> SessionReceiver
