+++ packages/@azure/servicebus/data-plane/lib/subscriptionClient.ts
@@ -3,11 +3,11 @@
 
 import * as log from "./log";
 import { ConnectionContext } from "./connectionContext";
-import { Receiver, MessageReceiverOptions, SessionReceiverOuter } from "./receiver";
+import { Receiver, MessageReceiverOptions, SessionReceiver } from "./receiver";
 import { ReceivedMessageInfo } from "./serviceBusMessage";
 import { Client } from "./client";
 import { CorrelationFilter, RuleDescription } from "./core/managementClient";
-import { SessionReceiver, SessionReceiverOptions } from "./session/messageSession";
+import { MessageSession, SessionReceiverOptions } from "./session/messageSession";
 
 export class SubscriptionClient extends Client {
   /**
@@ -202,11 +202,11 @@ export class SubscriptionClient extends Client {
    *
    * @returns SessionReceiver An instance of a SessionReceiver to receive messages from the session.
    */
-  async getSessionReceiver(options?: SessionReceiverOptions): Promise<SessionReceiverOuter> {
+  async getSessionReceiver(options?: SessionReceiverOptions): Promise<SessionReceiver> {
     if (!options) options = {};
     this._context.isSessionEnabled = true;
-    const sessionReceiver = await SessionReceiver.create(this._context, options);
-    return new SessionReceiverOuter(this._context, sessionReceiver);
+    const sessionReceiver = await MessageSession.create(this._context, options);
+    return new SessionReceiver(this._context, sessionReceiver);
   }
 
   //#endregion
