+++ packages/@azure/servicebus/data-plane/test/batchReceiver.spec.ts
@@ -26,7 +26,7 @@ import {
   ClientType,
   purge
 } from "./testUtils";
-import { Receiver, SessionReceiverOuter } from "../lib/receiver";
+import { Receiver, SessionReceiver } from "../lib/receiver";
 import { Sender } from "../lib/sender";
 
 async function testPeekMsgsLength(
@@ -49,7 +49,7 @@ let senderClient: QueueClient | TopicClient;
 let receiverClient: QueueClient | SubscriptionClient;
 let deadLetterClient: QueueClient | SubscriptionClient;
 let sender: Sender;
-let receiver: Receiver | SessionReceiverOuter;
+let receiver: Receiver | SessionReceiver;
 const maxDeliveryCount = 10;
 
 async function beforeEachTest(
