+++ packages/@azure/servicebus/data-plane/test/receiveAndDeleteMode.spec.ts
@@ -30,7 +30,7 @@ import {
   purge
 } from "./testUtils";
 
-import { Receiver, SessionReceiverOuter } from "../lib/receiver";
+import { Receiver, SessionReceiver } from "../lib/receiver";
 import { Sender } from "../lib/sender";
 
 async function testPeekMsgsLength(
@@ -52,7 +52,7 @@ let errorWasThrown: boolean;
 let senderClient: QueueClient | TopicClient;
 let receiverClient: QueueClient | SubscriptionClient;
 let sender: Sender;
-let receiver: Receiver | SessionReceiverOuter;
+let receiver: Receiver | SessionReceiver;
 
 async function beforeEachTest(
   senderType: ClientType,
